refactor: extract VideoPlayer into decode/layout/playback modules
Split the monolithic video_player.c (2606 lines down to 2097) into three
focused internal modules, each with its own file and internal header:

- video_player_decode: FFmpeg init, sws context, RGBA buffer allocation,
  decode resource teardown, PTS rescaling, draining flag, renderer presence
- video_player_layout: line cache, I/O timing averages, queue depth, slow
  level detection, late window calculation, pixel mode labels
- video_player_playback: playback generation counters, frame delay, EOF flags,
  clock/target PTS helpers, frame queue length, peek, calc_delay_ms,
  should_drop_late_frame

No behavior changes. All tests pass (272/272 C + 16 Python).
